headless-server/requirements.txt
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=9.5.0
schedule>=1.2.0
python-daemon>=3.0.1
requests>=2.31.0
